api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pyjwt>=2.9.0
pydantic>=2.10.0
